task_34_DB/_02_simple_queries/requirements.txt
python-dotenv
pyodbc